render feature setup note directly from the Behaviours component
